fix: enable spinGlobe and rotateAnimation correctly with globe mode
