+++ packages/svelte-maplibre-storymap/src/lib/StoryMapChapter.svelte
@@ -1,11 +1,7 @@
 <script lang="ts">
-	import type { mapProjectionType, StoryMapChapterType, StoryMapTemplate } from '$lib/interfaces';
-	import type { StyleSpecification } from 'maplibre-gl';
+	import type { StoryMapChapterType, StoryMapTemplate } from '$lib/interfaces';
 	import { marked } from 'marked';
-	import { getContext, untrack } from 'svelte';
-	import { layerTypes } from './helpers';
-	import { STORYMAP_MAPSTORE_CONTEXT_KEY, type MapStore } from './stores/map';
-	import { STORYMAP_MAPSTYLE_STORE_CONTEXT_KEY, type MapStyleStore } from './stores/mapStyle';
+	import { getContext } from 'svelte';
 	import {
 		STORYMAP_CONFIG_STORE_CONTEXT_KEY,
 		type StoryMapConfigStore
@@ -26,118 +22,9 @@
 	}: Props = $props();
 
 	// stores should be set at the parent component
-	let mapStore: MapStore = getContext(STORYMAP_MAPSTORE_CONTEXT_KEY);
-	let mapStyleStore: MapStyleStore = getContext(STORYMAP_MAPSTYLE_STORE_CONTEXT_KEY);
 	let config: StoryMapConfigStore = getContext(STORYMAP_CONFIG_STORE_CONTEXT_KEY);
 
 	let isLast = $derived($config.chapters[$config.chapters.length - 1]?.id === chapter.id);
-
-	const setChapterConfig = async () => {
-		if (!$mapStore) return;
-		if (!chapter) return;
-
-		if (chapter.id !== activeId) return;
-		if (chapter.style) {
-			if ($mapStyleStore !== chapter.style) {
-				$mapStyleStore = chapter.style;
-			}
-		} else if ($mapStyleStore !== $config.style) {
-			$mapStyleStore = $config.style;
-		}
-
-		if (typeof $mapStyleStore === 'string') {
-			const res = await fetch($mapStyleStore);
-			$mapStyleStore = await res.json();
-		}
-
-		const eventLength = chapter.onChapterEnter?.length ?? 0;
-		if (eventLength > 0) {
-			const newStyle: StyleSpecification = JSON.parse(JSON.stringify($mapStyleStore));
-			chapter.onChapterEnter?.forEach((layer) => {
-				const index = newStyle.layers.findIndex((l) => l.id === layer.layer);
-				if (index === -1) return;
-				const l = newStyle.layers[index];
-				const props = layerTypes[l.type];
-				if (props && props.length > 0) {
-					props.forEach((prop) => {
-						newStyle.layers[index].paint[prop] = layer.opacity;
-					});
-				} else {
-					const visibility = layer.opacity === 0 ? 'none' : 'visible';
-					if (!newStyle.layers[index].layout) {
-						newStyle.layers[index].layout = {};
-					}
-					newStyle.layers[index].layout.visibility = visibility;
-				}
-			});
-			$mapStyleStore = newStyle;
-		}
-
-		let mapProjection: mapProjectionType;
-		if (chapter.projection) {
-			mapProjection = chapter.projection;
-		} else if ($config.projection) {
-			mapProjection = $config.projection;
-		} else {
-			mapProjection = 'mercator';
-		}
-		if (
-			!(mapProjection && ($mapStyleStore as StyleSpecification).projection?.type === mapProjection)
-		) {
-			($mapStyleStore as StyleSpecification).projection = { type: mapProjection };
-		}
-
-		$mapStore.setStyle($mapStyleStore);
-
-		$mapStore[chapter.mapAnimation || 'flyTo']({
-			center: chapter.location.center,
-			zoom: chapter.location.zoom,
-			bearing: chapter.location.bearing ?? 0,
-			pitch: chapter.location.pitch ?? 0
-		});
-
-		if (chapter.mapInteractive) {
-			$mapStore.scrollZoom.disable(); //disable scrollZoom because it will conflict with scrolling chapters
-			$mapStore.boxZoom.enable();
-			$mapStore.dragRotate.enable();
-			$mapStore.dragPan.enable();
-			$mapStore.keyboard.enable();
-			$mapStore.doubleClickZoom.enable();
-			$mapStore.touchZoomRotate.enable();
-			$mapStore.touchPitch.enable();
-			$mapStore.getCanvas().style.cursor = 'grab';
-		} else {
-			$mapStore.scrollZoom.disable();
-			$mapStore.boxZoom.disable();
-			$mapStore.dragRotate.disable();
-			$mapStore.dragPan.disable();
-			$mapStore.keyboard.disable();
-			$mapStore.doubleClickZoom.disable();
-			$mapStore.touchZoomRotate.disable();
-			$mapStore.touchPitch.disable();
-			$mapStore.getCanvas().style.cursor = 'default';
-		}
-
-		if (chapter.rotateAnimation) {
-			$mapStore.once('moveend', () => {
-				const rotateNumber = $mapStore.getBearing();
-				$mapStore.rotateTo(rotateNumber + 180, {
-					duration: 30000,
-					easing: function (t) {
-						return t;
-					}
-				});
-			});
-		}
-	};
-
-	$effect(() => {
-		if (activeId !== undefined) {
-			untrack(() => {
-				setChapterConfig();
-			});
-		}
-	});
 </script>
 
 <section
